+++ examples/generic-colors/src/utils/index.ts
@@ -1,5 +1,5 @@
 import { ethers } from "ethers";
 
-const decodeColor = (color: string) => `#${ethers.utils.hexStripZeros(color).substr(2).toUpperCase()}`
+const decodeColor = (color: string) => `#${ethers.utils.hexStripZeros(color).substring(2).toUpperCase()}`
 
 export { decodeColor }