+++ packages/sdk/src/fee/feeOracle.ts
@@ -1,7 +1,6 @@
 import { FeeHandlerWithOracle__factory } from '@buildwithsygma/sygma-contracts';
-import { ethers } from 'ethers';
+import { ethers, utils } from 'ethers';
 import fetch from 'node-fetch';
-import EthCrypto from 'eth-crypto';
 
 import { OracleResource, FeeDataResult } from '../types';
 import { toHex, createERCDepositData } from '../utils/helpers';
@@ -17,15 +16,10 @@ type OracleResponse = {
  * @param oracleResponse - OracleResource object
  * @param amount - amount to send
  * @param tokenResource - the resource Id of the token
- * @param oraclePrivateKey - private key of the oracle
  * @returns {string} - hex of the fee data constructed
  */
-export const createOracleFeeData = (
-  oracleResponse: OracleResource,
-  amount: number,
-  tokenResource: string,
-  oraclePrivateKey?: string,
-): string => {
+
+export const createOracleFeeData = (oracleResponse: OracleResource, amount: string): string => {
   /*
         feeData structure:
             ber*10^18:    uint256
@@ -62,16 +56,8 @@ export const createOracleFeeData = (
     ],
   );
 
-  let signature;
-  if (oraclePrivateKey) {
-    // temprorary signature generated with sign by private key
-    const messageHash = EthCrypto.hash.keccak256([{ type: 'bytes', value: oracleMessage }]);
-    signature = EthCrypto.sign(oraclePrivateKey, messageHash);
-    return oracleMessage + signature.substr(2) + toHex(amount, 32).substr(2);
-  } else {
-    signature = oracleResponse.signature;
-    return oracleMessage + signature + toHex(0, 32).substr(2);
-  }
+  const signature = oracleResponse.signature;
+  return oracleMessage + signature + toHex(amount, 32).substring(2);
 };
 
 /**
@@ -90,8 +76,6 @@ export const calculateFeeData = async ({
   tokenAmount,
   feeOracleBaseUrl,
   feeOracleHandlerAddress,
-  overridedResourceId,
-  oraclePrivateKey,
 }: {
   provider: ethers.providers.Provider;
   sender: string;
@@ -99,31 +83,23 @@ export const calculateFeeData = async ({
   fromDomainID: number;
   toDomainID: number;
   resourceID: string;
-  tokenAmount: number;
+  tokenAmount: string;
   feeOracleBaseUrl: string;
   feeOracleHandlerAddress: string;
-  overridedResourceId?: string;
-  oraclePrivateKey?: string;
 }): Promise<FeeDataResult | undefined> => {
-  const depositData = createERCDepositData(tokenAmount, 20, recipientAddress);
+  const depositData = createERCDepositData(utils.parseUnits(tokenAmount, 18), 20, recipientAddress);
   let oracleResponse;
   try {
     oracleResponse = await requestFeeFromFeeOracle({
       feeOracleBaseUrl,
       fromDomainID,
       toDomainID,
-      resourceID: overridedResourceId ? overridedResourceId : resourceID, // dirty hack for localsetup
+      resourceID,
     });
   } catch (e) {
-    //@ts-ignore-line
-    return e;
+    return Promise.reject(e);
   }
-  const feeData = createOracleFeeData(
-    oracleResponse as OracleResource,
-    tokenAmount,
-    resourceID,
-    oraclePrivateKey,
-  );
+  const feeData = createOracleFeeData(oracleResponse as OracleResource, tokenAmount);
   const FeeHandlerWithOracleInstance = FeeHandlerWithOracle__factory.connect(
     feeOracleHandlerAddress,
     provider,
@@ -167,7 +143,7 @@ export const requestFeeFromFeeOracle = async ({
 }): Promise<OracleResource | undefined> => {
   try {
     const response = await fetch(
-      `${feeOracleBaseUrl}/v1/rate/from/${fromDomainID}/to/${toDomainID}/resourceid/${resourceID}/gasLimit/${msgGasLimit}`,
+      `${feeOracleBaseUrl}/v1/rate/from/${fromDomainID}/to/${toDomainID}/resourceid/${resourceID}?gasLimit=${msgGasLimit}`,
       {
         headers: {
           'Cache-Control': 'no-cache',
@@ -185,6 +161,6 @@ export const requestFeeFromFeeOracle = async ({
       return data.response;
     }
   } catch (e) {
-    return Promise.reject(new Error('Invalid fee oracle response'));
+    return Promise.reject(e);
   }
 };